backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# MongoDB
motor==3.3.2  # 异步MongoDB驱动
pymongo==4.6.1
beanie==1.24.0  # MongoDB ODM

# AI/ML
langchain==0.1.0
langchain-openai==0.0.2
openai==1.8.0
tiktoken==0.5.2
numpy==1.26.3
tenacity==8.2.3  # 重试机制

# Azure
azure-storage-blob==12.19.0
azure-identity==1.15.0

# File Processing
aiofiles==23.2.1
python-magic==0.4.27
chardet==5.2.0

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0